Move CircularBufferConfigImpl to impl/
