+++ tests/tt_metal/tt_metal/api/circular_buffer/test_CircularBuffer_allocation.cpp
@@ -37,6 +37,7 @@
 // Access to internal API: ProgramImpl::get_cb_base_addr
 #include "impl/program/program_impl.hpp"
 #include "impl/buffers/circular_buffer_constants.h"
+#include "impl/buffers/circular_buffer_config.hpp"
 
 using std::vector;
 using namespace tt::tt_metal;
